+++ modules/astra/module_lua.h
@@ -120,7 +120,7 @@ typedef struct
     }
 
 bool module_option_number(const char *name, int *number);
-bool module_option_string(const char *name, const char **string, int *length);
+bool module_option_string(const char *name, const char **string, size_t *length);
 bool module_option_boolean(const char *name, bool *boolean);
 
 #endif /* _MODULE_LUA_H_ */
